// src/app/core/models/movie.ts
export interface Movie {
  adult: boolean
  backdrop_path: string
  id: number
  title: string
  original_title: string
  overview: string
  poster_path: string
  media_type: string
  original_language: string
  genre_ids: number[]
  popularity: number
  release_date: string
  video: boolean
  vote_average: number
  vote_count: number
}

export interface MovieCollection {
  page: number,
  results: Movie[],
  total_pages: number,
  total_results: number,
}